+++ src/roiextractors/extractors/schnitzerextractor/cnmfesegmentationextractor.py
@@ -13,7 +13,11 @@ import numpy as np
 from lazy_ops import DatasetView
 
 from ...extraction_tools import PathType
-from ...segmentationextractor import RoiResponse, SegmentationExtractor
+from ...segmentationextractor import (
+    RoiResponse,
+    SegmentationExtractor,
+    _ROIMasks,
+)
 
 
 class CnmfeSegmentationExtractor(SegmentationExtractor):
@@ -42,11 +46,24 @@ class CnmfeSegmentationExtractor(SegmentationExtractor):
         SegmentationExtractor.__init__(self)
         self.file_path = file_path
         self._dataset_file, self._group0 = self._file_extractor_read()
-        self._image_masks = self._image_mask_extractor_read()
+
+        # Read traces first to get number of ROIs
         traces = self._trace_extractor_read()
         cell_ids = list(range(traces.shape[1]))
         self._roi_ids = cell_ids
         self._roi_responses.append(RoiResponse("raw", traces, cell_ids))
+
+        # Create ROI representations from dense image masks
+        image_masks_data = self._image_mask_extractor_read()  # DatasetView (H, W, N)
+        roi_id_map = {roi_id: index for index, roi_id in enumerate(cell_ids)}
+
+        self._roi_masks = _ROIMasks(
+            data=image_masks_data,
+            mask_tpe="nwb-image_mask",
+            field_of_view_shape=self.get_frame_shape(),
+            roi_id_map=roi_id_map,
+        )
+
         self._raw_movie_file_location = self._raw_datafile_read()
         self._sampling_frequency = self.get_num_frames() / self._tot_exptime_extractor_read()
         # self._sampling_frequency = self._dataset_file[self._group0[0]]['inputOptions']["Fs"][...][0][0]